+++ src/gui/mod.rs
@@ -114,8 +114,7 @@ pub enum Message {
     BackupFinished(Result<NaiveDateTime>),
     CatalogDownloaded(Result<Catalog>),
     CatalogInstallAddonFetched((Flavor, u32, Result<Addon>)),
-    FetchedCurseChangelog((Addon, AddonVersionKey, Result<(String, String)>)),
-    FetchedTukuiChangelog((Addon, AddonVersionKey, Result<(String, String)>)),
+    FetchedChangelog((Addon, AddonVersionKey, Result<(String, String)>)),
     AjourUpdateDownloaded(Result<(String, PathBuf)>),
 }
 
@@ -657,6 +656,7 @@ pub enum ColumnKey {
     Author,
     GameVersion,
     DateReleased,
+    Source,
 }
 
 impl ColumnKey {
@@ -672,6 +672,7 @@ impl ColumnKey {
             Author => "Author",
             GameVersion => "Game Version",
             DateReleased => "Latest Release",
+            Source => "Source",
         };
 
         title.to_string()
@@ -689,6 +690,7 @@ impl ColumnKey {
             Author => "author",
             GameVersion => "game_version",
             DateReleased => "date_released",
+            Source => "source",
         };
 
         s.to_string()
@@ -706,6 +708,7 @@ impl From<&str> for ColumnKey {
             "author" => ColumnKey::Author,
             "game_version" => ColumnKey::GameVersion,
             "date_released" => ColumnKey::DateReleased,
+            "source" => ColumnKey::Source,
             _ => panic!(format!("Unknown ColumnKey for {}", s)),
         }
     }
@@ -805,6 +808,13 @@ impl Default for HeaderState {
                     hidden: true,
                     order: 7,
                 },
+                ColumnState {
+                    key: ColumnKey::Source,
+                    btn_state: Default::default(),
+                    width: Length::Units(110),
+                    hidden: true,
+                    order: 8,
+                },
             ],
         }
     }
@@ -890,6 +900,12 @@ impl Default for ColumnSettings {
                 ColumnSettingState {
                     key: ColumnKey::DateReleased,
                     order: 7,
+                    up_btn_state: Default::default(),
+                    down_btn_state: Default::default(),
+                },
+                ColumnSettingState {
+                    key: ColumnKey::Source,
+                    order: 8,
                     up_btn_state: Default::default(),
                     down_btn_state: Default::default(),
                 },