+++ src/gui/update.rs
@@ -17,7 +17,7 @@ use {
         parse::{read_addon_directory, update_addon_fingerprint, FingerprintCollection},
         tukui_api,
         utility::{download_update_to_temp_file, wow_path_resolution},
-        Result,
+        wowi_api, Result,
     },
     async_std::sync::{Arc, Mutex},
     iced::{Command, Length},
@@ -332,10 +332,9 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
             let path = wow_path_resolution(chosen_path);
             log::debug!("Message::UpdateWowDirectory(Resolution({:?}))", &path);
 
-            // Clear addons.
-            ajour.addons = HashMap::new();
-
             if path.is_some() {
+                // Clear addons.
+                ajour.addons = HashMap::new();
                 // Update the path for World of Warcraft.
                 ajour.config.wow.directory = path;
                 // Persist the newly updated config.
@@ -440,7 +439,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                                     ExpandType::Changelog(Changelog::Loading(addon.clone(), *key));
                                 return Ok(Command::perform(
                                     perform_fetch_curse_changelog(addon.clone(), *key, id, file_id),
-                                    Message::FetchedCurseChangelog,
+                                    Message::FetchedChangelog,
                                 ));
                             }
                         }
@@ -457,7 +456,19 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                                         ajour.config.wow.flavor,
                                         *key,
                                     ),
-                                    Message::FetchedTukuiChangelog,
+                                    Message::FetchedChangelog,
+                                ));
+                            }
+                        }
+
+                        // If we have a Wowi addon.
+                        if addon.active_repository == Some(Repository::WowI) {
+                            if let Some(id) = addon.repository_id() {
+                                ajour.expanded_type =
+                                    ExpandType::Changelog(Changelog::Loading(addon.clone(), *key));
+                                return Ok(Command::perform(
+                                    perform_fetch_wowi_changelog(addon.clone(), id, *key),
+                                    Message::FetchedChangelog,
                                 ));
                             }
                         }
@@ -1377,11 +1388,8 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                 );
             }
         },
-        Message::FetchedTukuiChangelog((addon, key, result)) => {
-            log::debug!(
-                "Message::FetchedTukuiChangelog(error: {})",
-                &result.is_err()
-            );
+        Message::FetchedChangelog((addon, key, result)) => {
+            log::debug!("Message::FetchedChangelog(error: {})", &result.is_err());
             match result {
                 Ok((changelog, url)) => {
                     let payload = ChangelogPayload { changelog, url };
@@ -1389,25 +1397,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                     ajour.expanded_type = ExpandType::Changelog(changelog);
                 }
                 Err(error) => {
-                    log::error!("Message::FetchedTukuiChangelog(error: {})", &error);
-                    ajour.expanded_type = ExpandType::None;
-                }
-            }
-        }
-        Message::FetchedCurseChangelog((addon, key, result)) => {
-            log::debug!(
-                "Message::FetchedCurseChangelog(error: {})",
-                &result.is_err()
-            );
-
-            match result {
-                Ok((changelog, url)) => {
-                    let payload = ChangelogPayload { changelog, url };
-                    let changelog = Changelog::Some(addon, payload, key);
-                    ajour.expanded_type = ExpandType::Changelog(changelog);
-                }
-                Err(error) => {
-                    log::error!("Message::FetchedCurseChangelog(error: {})", &error);
+                    log::error!("Message::FetchedChangelog(error: {})", &error);
                     ajour.expanded_type = ExpandType::None;
                 }
             }
@@ -1523,6 +1513,22 @@ async fn perform_fetch_curse_changelog(
     (addon, key, curse_api::fetch_changelog(id, file_id).await)
 }
 
+async fn perform_fetch_wowi_changelog(
+    addon: Addon,
+    id: String,
+    key: AddonVersionKey,
+) -> (Addon, AddonVersionKey, Result<(String, String)>) {
+    (
+        addon,
+        key,
+        Ok((
+            "Please view this changelog in the browser by pressing 'Full Changelog' to the right"
+                .to_owned(),
+            wowi_api::changelog_url(id),
+        )),
+    )
+}
+
 /// Downloads the newest version of the addon.
 /// This is for now only downloading from warcraftinterface.
 async fn perform_download_addon(
@@ -1589,14 +1595,18 @@ async fn perform_fetch_latest_addon(
 fn sort_addons(addons: &mut [Addon], sort_direction: SortDirection, column_key: ColumnKey) {
     match (column_key, sort_direction) {
         (ColumnKey::Title, SortDirection::Asc) => {
-            addons.sort();
+            addons.sort_by(|a, b| a.title().to_lowercase().cmp(&b.title().to_lowercase()));
         }
         (ColumnKey::Title, SortDirection::Desc) => {
             addons.sort_by(|a, b| {
-                a.title().cmp(&b.title()).reverse().then_with(|| {
-                    a.relevant_release_package()
-                        .cmp(&b.relevant_release_package())
-                })
+                a.title()
+                    .to_lowercase()
+                    .cmp(&b.title().to_lowercase())
+                    .reverse()
+                    .then_with(|| {
+                        a.relevant_release_package()
+                            .cmp(&b.relevant_release_package())
+                    })
             });
         }
         (ColumnKey::LocalVersion, SortDirection::Asc) => {
@@ -1672,6 +1682,12 @@ fn sort_addons(addons: &mut [Addon], sort_direction: SortDirection, column_key: 
                     .cmp(&b.relevant_release_package().map(|p| p.date_time))
                     .reverse()
             });
+        }
+        (ColumnKey::Source, SortDirection::Asc) => {
+            addons.sort_by(|a, b| a.active_repository.cmp(&b.active_repository))
+        }
+        (ColumnKey::Source, SortDirection::Desc) => {
+            addons.sort_by(|a, b| a.active_repository.cmp(&b.active_repository).reverse())
         }
     }
 }